Tilemap for hieroglyphs.
we decreased draw call in demo level for full screen view on 1080p from 200 draw calls to something like 130/140 when camera is on osiris at the beginning of the level.
LEVEL EDITOR NEW FEATURE: i've added a reload level button in the level editor to be able to automatically reload instantiated prefab, for some reason, when instanitating through the level editor we could not reverting to prefab anymore, so this allow us to make changes to the tiles prefabs and then apply them by recreating the level in one click.

I've also removed some old 3d-cube based tiles.

Finally i moved the waterarea collider in separate gameobjects, this allow us to reload water tiles without having to reset the water area again